+++ src/main/java/com/example/demo/service/general/MypageService.java
@@ -1,6 +1,7 @@
 package com.example.demo.service.general;
 
 import com.example.demo.domain.dto.MypageResponseDTO;
+import com.example.demo.domain.dto.Review.ReviewWithAverageDTO;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface MypageService {
@@ -8,4 +9,9 @@ public interface MypageService {
     MypageResponseDTO.ApplyListDto getApplies();
 
     String updateProfileImage(Long userId, MultipartFile file);
+
+    //내 리뷰 조회
+    ReviewWithAverageDTO getReviewsByUserId(Long userId);
+
+
 }
