+++ src/main/java/com/example/demo/service/general/ReviewService.java
@@ -16,9 +16,6 @@ public interface ReviewService {
     // 리뷰 삭제
     void deleteReview(Long reviewId, ReviewDeleteDTO reviewDelete);
 
-    //리뷰 조회
-    ReviewWithAverageDTO getReviewsByUserId(Long userId);
-
 }
